refactor: RecipeFood amount Double로 변경 및 칼로리 부분 주석 처리

+++ src/main/java/com/tablelog/tablelogback/domain/recipe_food/dto/controller/RecipeFoodCreateControllerRequestDto.java
@@ -3,7 +3,7 @@ package com.tablelog.tablelogback.domain.recipe_food.dto.controller;
 import com.tablelog.tablelogback.global.enums.FoodUnit;
 
 public record RecipeFoodCreateControllerRequestDto(
-        Integer amount,
+        Double amount,
         FoodUnit recipeFoodUnit,
         Long foodId
 ) {
